Rettet kompileringsfeil og kjørt ktlint
